Sprint 3 HackerEarth

+++ POS/src/main/java/com/ust/pos/api/user/UserControllerApi.java
@@ -3,13 +3,15 @@ package com.ust.pos.api.user;
 import com.ust.pos.api.BaseController;
 import com.ust.pos.dto.PaginationDto;
 import com.ust.pos.dto.UserDto;
+import com.ust.pos.dto.WsDto;
+import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/user")
@@ -18,19 +20,22 @@ public class UserControllerApi extends BaseController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private RoleService roleService;
+
     @PostMapping("/list")
-    public List<UserDto> list(@RequestBody PaginationDto paginationDto) {
+    public WsDto<UserDto> list(@RequestBody PaginationDto paginationDto) {
         Pageable pageable = getPageable(paginationDto.getPage(), paginationDto.getSizePerPage(), paginationDto.getSortDirection(), paginationDto.getSortField());
         return userService.findAll(pageable);
     }
 
     @PostMapping("/register")
-    public UserDto add(@RequestBody UserDto userDto) {
+    public UserDto add(@RequestBody UserDto userDto){
         return userService.save(userDto);
     }
 
     @GetMapping("/get")
-    public UserDto update(@RequestParam String username, @RequestBody UserDto userDto) {
+    public UserDto update(@RequestParam String username) {
         return userService.findByUserName(username);
     }
 
@@ -42,7 +47,17 @@ public class UserControllerApi extends BaseController {
     @GetMapping("/delete")
     public boolean delete(Model model, @RequestParam String username) {
         try {
-            userService.delete(username);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication != null) {
+                String loggedInUser = authentication.getName();
+                if (loggedInUser != null) {
+                    userService.delete(username);
+                    if (loggedInUser.equals(username)) {
+                        SecurityContextHolder.clearContext();
+                        return true;
+                    }
+                }
+            }
         } catch (Exception e) {
             return false;
         }
